+++ labfreed/__init__.py
@@ -5,8 +5,8 @@ Python implementation of LabFREED building blocks
 __version__ = "1.0.0"
 
 from labfreed.pac_id import *  # noqa: F403
-from labfreed.pac_cat import * # noqa: F403
-from labfreed.pac_id_resolver import * # noqa: F403
-from labfreed.trex import * # noqa: F403
-from labfreed.pac_attributes import * # noqa: F403
+from labfreed.pac_cat import *  # noqa: F403
+from labfreed.pac_id_resolver import *  # noqa: F403
+from labfreed.trex import *  # noqa: F403
+from labfreed.pac_attributes import *  # noqa: F403
 from labfreed.labfreed_infrastructure import *  # noqa: F403